Overall Wireframe
Made separate pages for each of the likely tabs we need to deal with.
Also linked them up. Now we can make the rest of a wireframe from it.

There will likely be a couple more states for each screen to fake
interactivity. We can decide on features in the future.